+++ nodalia-cards.manifest.js
@@ -1,11 +1,11 @@
 export default {
-  "pkgVersion": "2.2.1-beta.1",
-  "contentSha256_12": "8c1545d4c71a",
+  "pkgVersion": "2.2.1-beta.2",
+  "contentSha256_12": "eebf8e9c607f",
   "file": "nodalia-cards.js",
   "loaderFile": "nodalia-cards.js",
   "hacsFile": "nodalia-cards.js",
   "editorSha256_12": "fcf4f0088891"
 };
-export const pkgVersion = "2.2.1-beta.1";
-export const contentSha256_12 = "8c1545d4c71a";
+export const pkgVersion = "2.2.1-beta.2";
+export const contentSha256_12 = "eebf8e9c607f";
 export const file = "nodalia-cards.js";